created POST route and refactored mongoose
