--- mongodb.js
@@ -1,182 +1,0 @@
-// CRUD Create Read Update Delete
-
-// Server downloaded on machine
-// See it in /Users/hughej/mongodb-data
-// Command for initialising the database
-// /Users/hughej/mongodb/bin/mongod --dbpath=/Users/hughej/mongodb-data/
-
-// Client needed to intialise connection to MongoDB
-// const mongodb = require('mongodb');
-// const MongoClient = mongodb.MongoClient;
-// const ObjectID = mongodb = mongodb.ObjectID;
-
-const { MongoClient, ObjectID } = require('mongodb');
-
-// Full IP of localhost used to prevent weird localhost issues
-const connectionUrl = 'mongodb://127.0.0.1:27017';
-const databaseName = 'task-manager';
-
-MongoClient.connect(connectionUrl, { useNewUrlParser: true }, (error, client) => {
-    if (error) {
-        return console.log('Unable to connect to database!')
-    }
-
-    const db = client.db(databaseName);
-
-    // CREATE
-
-    // db.collection('users').insertOne({
-    //     _id: id,
-    //     name: 'Wayne',
-    //     age: 40,
-    // }, (error, result) => {
-    //     if (error) {
-    //         return console.log('Unable to insert user')
-    //     }
-
-    //     console.log(result.ops);
-    // })
-
-    // db.collection('users').insertMany([
-    //     {
-    //         name: 'Caryn',
-    //         age: 39
-    //     },
-    //     {
-    //         name: 'James',
-    //         age: 36
-    //     }
-    // ], (error, result) => {
-    //     if (error) {
-    //         return console.log('Unable to insert documents!')
-    //     }
-
-    //     console.log(result.ops)
-    // })
-
-    // db.collection('tasks').insertMany([
-    //     {
-    //         James: 'empty bin',
-    //         James: 'make bed'
-    //     },
-    //     {
-    //         Caryn: 'order groceries',
-    //         Caryn: 'Make dinner'
-    //     },
-    //     {
-    //         Gabriel: 'eat breakfast',
-    //         Gabriel: 'put on cape and shield'
-    //     }
-    // ], (error, result) => {
-    //     if (error) {
-    //         return console.log('Unable to insert tasks')
-    //     }
-
-    //     console.log(result.ops);
-    // })
-// })
-
-    // READ
-
-    // db.collection('users').findOne({_id: new ObjectID("5ccae7cb45a8893b5b9821ff") }, (error, user) => {
-    //     if (error) {
-    //         return console.log('Unable to fetch')
-    //     }
-    //     console.log(user);
-    // })
-
-    // find returns a cursor not a callback
-    // which is a pointer to the data in the db
-    // console log array of users matching age: 36
-    // db.collection('users').find({age: 36}).toArray((error, users) => {
-    //     console.log(users);
-    // });
-    
-    // same as above but console logging count matching age: 36
-    // db.collection('users').find({age: 36}).count((error, count) => {
-    //     console.log(count);
-    // });
-
-    // db.collection('tasks').insertMany([
-    //     {
-    //         name: 'Dishes',
-    //         completed: false
-    //     },
-    //     {
-    //         name: 'Life Insurance',
-    //         completed: false
-    //     },
-    //     {
-    //         name: 'DrumKittutorial',
-    //         completed: true
-    //     }
-    // ], (error, result) => {
-    //     if (error) {
-    //         return console.log('Unable to insert documents!')
-    //     }
-    //     console.log(result.ops);
-    // })
-
-    // db.collection('tasks').findOne({_id: new ObjectID('5ccaee913f4c4a404c08e592')}, (error, task) => {
-    //     if (error) {
-    //         return console.log('Unable to fetch')
-    //     }
-    //     console.log(task);
-    // })
-
-    // db.collection('tasks').find({completed: false}).toArray((error, tasks) => {
-    //     if (error) {
-    //         return console.log('Cannot find tasks')
-    //     }
-    //     console.log(tasks);
-    // })
-
-    // UPDATE
-
-    // updateOne returns a promise
-    // db.collection('users').updateOne({
-    //     _id: new ObjectID('5ccae7cb45a8893b5b9821ff'),    
-    // }, {
-    //     $inc: {
-    //         age: -6
-    //     }
-    // }).then((result) => {
-    //     console.log(result)
-    // }).catch((error) => {
-    //     console.log(error)
-    // })
-
-    // db.collection('tasks').updateMany({
-    //     completed: false
-    // }, {
-    //     $set: {
-    //         completed: true
-    //     }
-    // }).then((result) => {
-    //     console.log(result)
-    // }).catch((error) => {
-    //     console.log(error)
-    // })
-
-    // DELETE
-
-    db.collection('tasks').deleteOne(
-        {
-            name: 'DrumKittutorial' 
-        }
-    ).then((result) => {
-        console.log(result)
-    }).catch((error) => {
-        console.log(error)
-    })
-
-    db.collection('users').deleteMany(
-            {
-                name: 'James'
-            },
-    ).then((result) => {
-            console.log(result)
-        }).catch((error) => {
-            console.log(error)
-        })
-})
